// fhf-datafangst-client/src/generated/openapi/models/landing.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Gear } from './gear';
// May contain unused imports in some cases
// @ts-ignore
import type { GearGroup } from './gear-group';
// May contain unused imports in some cases
// @ts-ignore
import type { LandingCatch } from './landing-catch';
// May contain unused imports in some cases
// @ts-ignore
import type { VesselLengthGroup } from './vessel-length-group';

/**
 * 
 * @export
 * @interface Landing
 */
export interface Landing {
    /**
     * 
     * @type {string}
     * @memberof Landing
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof Landing
     */
    'tripId'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Landing
     */
    'landingTimestamp': string;
    /**
     * 
     * @type {string}
     * @memberof Landing
     */
    'catchLocation'?: string | null;
    /**
     * 
     * @type {Gear}
     * @memberof Landing
     */
    'gearId': Gear;
    /**
     * 
     * @type {GearGroup}
     * @memberof Landing
     */
    'gearGroupId': GearGroup;
    /**
     * 
     * @type {string}
     * @memberof Landing
     */
    'deliveryPointId'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Landing
     */
    'fiskeridirVesselId'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Landing
     */
    'vesselCallSign'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Landing
     */
    'vesselName'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Landing
     */
    'vesselLength'?: number | null;
    /**
     * 
     * @type {VesselLengthGroup}
     * @memberof Landing
     */
    'vesselLengthGroup': VesselLengthGroup;
    /**
     * 
     * @type {number}
     * @memberof Landing
     */
    'totalLivingWeight': number;
    /**
     * 
     * @type {number}
     * @memberof Landing
     */
    'totalProductWeight': number;
    /**
     * 
     * @type {number}
     * @memberof Landing
     */
    'totalGrossWeight': number;
    /**
     * 
     * @type {Array<LandingCatch>}
     * @memberof Landing
     */
    'catches': Array<LandingCatch>;
}